changed add task from file to json string

+++ GridCalculator/src/main/java/ru/paramonova/dto/Result.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Builder
 @ToString
 public class Result {
-    int batchId;
+    long batchId;
     boolean connected;
     List<Pipe> pipes;
     List<Line> lines;
